make the profile page pass url params into tabbed viewer and suchnot
